Move tests to dedicated folder
This commit moves the tests from lib.rs to their own tests directory and as a side effect reformats some of the code
